+++ tests/RdKafka/ConfTest.php
@@ -160,9 +160,6 @@ class ConfTest extends TestCase
         $conf->get('unknown.property');
     }
 
-    /**
-     * @small
-     */
     public function testSetLogCb(): void
     {
         $loggerCallbacks = 0;
@@ -184,9 +181,6 @@ class ConfTest extends TestCase
         $this->assertSame(1, $loggerCallbacks, 'Expected debug level log callback');
     }
 
-    /**
-     * @small
-     */
     public function testSetErrorCb(): void
     {
         $errorCallbackStack = [];
@@ -226,7 +220,7 @@ class ConfTest extends TestCase
         $producerTopic = $producer->newTopic(KAFKA_TEST_TOPIC);
         $producerTopic->produce(0, 0, __METHOD__ . '1');
         $producerTopic->produce(0, 0, __METHOD__ . '2');
-        $producer->poll((int) KAFKA_TEST_TIMEOUT_MS);
+        $producer->poll(KAFKA_TEST_TIMEOUT_MS);
 
         $this->assertCount(2, $drMsgCallbackStack);
         $this->assertSame($producer, $drMsgCallbackStack[0]['producer']);
@@ -234,12 +228,9 @@ class ConfTest extends TestCase
         $this->assertSame($producer, $drMsgCallbackStack[1]['producer']);
         $this->assertSame(__METHOD__ . '2', $drMsgCallbackStack[1]['message']->payload);
 
-        $producer->flush((int) KAFKA_TEST_TIMEOUT_MS);
+        $producer->flush(KAFKA_TEST_TIMEOUT_MS);
     }
 
-    /**
-     * @small
-     */
     public function testSetStatsCb(): void
     {
         $statsJson = '';
@@ -265,9 +256,6 @@ class ConfTest extends TestCase
         $this->assertSame([], $stats['brokers']);
     }
 
-    /**
-     * @medium
-     */
     public function testSetRebalanceCb(): void
     {
         $rebalanceCallbackStack = [];
