+++ tests/RdKafka/ConsumerTopicTest.php
@@ -48,14 +48,14 @@ class ConsumerTopicTest extends TestCase
         $producer->addBrokers(KAFKA_BROKERS);
         $producerTopic = $producer->newTopic(KAFKA_TEST_TOPIC);
         $producerTopic->produce(0, 0, __METHOD__);
-        $producer->flush((int) KAFKA_TEST_TIMEOUT_MS);
+        $producer->flush(KAFKA_TEST_TIMEOUT_MS);
 
         $consumer = new Consumer();
         $consumer->addBrokers(KAFKA_BROKERS);
         $consumerTopic = $consumer->newTopic(KAFKA_TEST_TOPIC);
         $consumerTopic->consumeStart(0, rd_kafka_offset_tail(1));
 
-        $message = $consumerTopic->consume(0, (int) KAFKA_TEST_TIMEOUT_MS);
+        $message = $consumerTopic->consume(0, KAFKA_TEST_TIMEOUT_MS);
 
         $consumerTopic->consumeStop(0);
 
@@ -70,7 +70,7 @@ class ConsumerTopicTest extends TestCase
 
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessageMatches('/partition/');
-        $consumerTopic->consume(-2, (int) KAFKA_TEST_TIMEOUT_MS);
+        $consumerTopic->consume(-2, KAFKA_TEST_TIMEOUT_MS);
     }
 
     public function testConsumeCallback(): void
@@ -84,7 +84,7 @@ class ConsumerTopicTest extends TestCase
         $producer = new Producer($conf);
         $producerTopic = $producer->newTopic(KAFKA_TEST_TOPIC);
         $producerTopic->produce(0, 0, __METHOD__);
-        $producer->flush((int) KAFKA_TEST_TIMEOUT_MS);
+        $producer->flush(KAFKA_TEST_TIMEOUT_MS);
 
         $consumer = new Consumer($conf);
         $consumerTopic = $consumer->newTopic(KAFKA_TEST_TOPIC);
@@ -93,7 +93,7 @@ class ConsumerTopicTest extends TestCase
         $callback = function (Message $message, $opaque = null) use (&$consumedMessage): void {
             $consumedMessage = $message;
         };
-        $messagesConsumed = $consumerTopic->consumeCallback(0, (int) KAFKA_TEST_TIMEOUT_MS, $callback);
+        $messagesConsumed = $consumerTopic->consumeCallback(0, KAFKA_TEST_TIMEOUT_MS, $callback);
 
         $consumerTopic->consumeStop(0);
 
@@ -111,14 +111,14 @@ class ConsumerTopicTest extends TestCase
         for ($i = 0; $i < $batchSize; $i++) {
             $producerTopic->produce(0, 0, __METHOD__ . $i);
         }
-        $producer->flush((int) KAFKA_TEST_TIMEOUT_MS);
+        $producer->flush(KAFKA_TEST_TIMEOUT_MS);
 
         $consumer = new Consumer();
         $consumer->addBrokers(KAFKA_BROKERS);
         $consumerTopic = $consumer->newTopic(KAFKA_TEST_TOPIC);
         $consumerTopic->consumeStart(0, rd_kafka_offset_tail($batchSize));
 
-        $messages = $consumerTopic->consumeBatch(0, (int) KAFKA_TEST_TIMEOUT_MS, $batchSize);
+        $messages = $consumerTopic->consumeBatch(0, KAFKA_TEST_TIMEOUT_MS, $batchSize);
 
         $consumerTopic->consumeStop(0);
 
@@ -139,7 +139,7 @@ class ConsumerTopicTest extends TestCase
 
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessageMatches('/partition/');
-        $consumerTopic->consumeBatch(-2, (int) KAFKA_TEST_TIMEOUT_MS, 1);
+        $consumerTopic->consumeBatch(-2, KAFKA_TEST_TIMEOUT_MS, 1);
     }
 
     public function testConsumeBatchWithInvalidBatchSizeShouldFail(): void
@@ -150,7 +150,7 @@ class ConsumerTopicTest extends TestCase
 
         $this->expectException(InvalidArgumentException::class);
         $this->expectExceptionMessageMatches('/batch_size/');
-        $consumerTopic->consumeBatch(0, (int) KAFKA_TEST_TIMEOUT_MS, 0);
+        $consumerTopic->consumeBatch(0, KAFKA_TEST_TIMEOUT_MS, 0);
     }
 
     public function testConsumeQueueStart(): void
@@ -159,7 +159,7 @@ class ConsumerTopicTest extends TestCase
         $producer->addBrokers(KAFKA_BROKERS);
         $producerTopic = $producer->newTopic(KAFKA_TEST_TOPIC);
         $producerTopic->produce(0, 0, __METHOD__);
-        $producer->flush((int) KAFKA_TEST_TIMEOUT_MS);
+        $producer->flush(KAFKA_TEST_TIMEOUT_MS);
 
         $consumer = new Consumer();
         $consumer->addBrokers(KAFKA_BROKERS);
@@ -167,7 +167,7 @@ class ConsumerTopicTest extends TestCase
         $consumerTopic = $consumer->newTopic(KAFKA_TEST_TOPIC);
 
         $consumerTopic->consumeQueueStart(0, rd_kafka_offset_tail(1), $queue);
-        $message = $queue->consume((int) KAFKA_TEST_TIMEOUT_MS);
+        $message = $queue->consume(KAFKA_TEST_TIMEOUT_MS);
         $consumerTopic->consumeStop(0);
 
         $this->assertSame(__METHOD__, $message->payload);
@@ -198,7 +198,7 @@ class ConsumerTopicTest extends TestCase
 
         $topicPartitions = $highLevelConsumer->getCommittedOffsets(
             [new TopicPartition(KAFKA_TEST_TOPIC, 0)],
-            (int) KAFKA_TEST_TIMEOUT_MS
+            KAFKA_TEST_TIMEOUT_MS
         );
         $this->assertSame(-1001, $topicPartitions[0]->getOffset());
 
@@ -214,7 +214,7 @@ class ConsumerTopicTest extends TestCase
 
         $topicPartitions = $highLevelConsumer->getCommittedOffsets(
             [new TopicPartition(KAFKA_TEST_TOPIC, 0)],
-            (int) KAFKA_TEST_TIMEOUT_MS
+            KAFKA_TEST_TIMEOUT_MS
         );
         $this->assertSame(5, $topicPartitions[0]->getOffset());
     }
