+++ tests/RdKafka/KafkaConsumerTest.php
@@ -25,7 +25,7 @@ class KafkaConsumerTest extends TestCase
         $producerTopic = $producer->newTopic(KAFKA_TEST_TOPIC);
         $producerTopic->produce(RD_KAFKA_PARTITION_UA, 0, 'payload-kafka-consumer-1');
         $producerTopic->produce(RD_KAFKA_PARTITION_UA, 0, 'payload-kafka-consumer-2');
-        $producer->flush((int) KAFKA_TEST_TIMEOUT_MS);
+        $producer->flush(KAFKA_TEST_TIMEOUT_MS);
     }
 
     public function testConstructWithMissingGroupIdConfShouldFail(): void
@@ -119,9 +119,6 @@ class KafkaConsumerTest extends TestCase
         $this->assertSame(KAFKA_TEST_TOPIC, $topicPartitions[0]);
     }
 
-    /**
-     * @medium
-     */
     public function testConsume(): void
     {
         $conf = new Conf();
@@ -137,7 +134,7 @@ class KafkaConsumerTest extends TestCase
 
         $lastMessage = $message = null;
         while (true) {
-            $message = $consumer->consume((int) KAFKA_TEST_TIMEOUT_MS);
+            $message = $consumer->consume(KAFKA_TEST_TIMEOUT_MS);
             if ($message->err === RD_KAFKA_RESP_ERR__TIMED_OUT) {
                 if ($lastMessage === null) {
                     continue;
@@ -185,9 +182,6 @@ class KafkaConsumerTest extends TestCase
         $consumer->commitAsync(new stdClass());
     }
 
-    /**
-     * @medium
-     */
     public function testCommitWithMessage(): void
     {
         $conf = new Conf();
@@ -212,7 +206,7 @@ class KafkaConsumerTest extends TestCase
 
         $lastMessage = $message = null;
         while (true) {
-            $message = $consumer->consume((int) KAFKA_TEST_TIMEOUT_MS);
+            $message = $consumer->consume(KAFKA_TEST_TIMEOUT_MS);
             if ($message->err === RD_KAFKA_RESP_ERR__TIMED_OUT) {
                 if ($lastMessage === null) {
                     continue;
@@ -225,7 +219,7 @@ class KafkaConsumerTest extends TestCase
         $consumer->commit($lastMessage);
 
         // just trigger callback
-        $consumer->consume((int) KAFKA_TEST_TIMEOUT_MS);
+        $consumer->consume(KAFKA_TEST_TIMEOUT_MS);
 
         $this->assertSame($message->offset + 1, $offset);
 
@@ -252,7 +246,7 @@ class KafkaConsumerTest extends TestCase
             [
                 new TopicPartition(KAFKA_TEST_TOPIC, 0),
             ],
-            (int) KAFKA_TEST_TIMEOUT_MS
+            KAFKA_TEST_TIMEOUT_MS
         );
 
         $this->assertCount(1, $topicPartitions);
@@ -313,7 +307,7 @@ class KafkaConsumerTest extends TestCase
             [
                 new TopicPartition(KAFKA_TEST_TOPIC, 0),
             ],
-            (int) KAFKA_TEST_TIMEOUT_MS
+            KAFKA_TEST_TIMEOUT_MS
         );
 
         $this->assertCount(1, $topicPartitions);
@@ -339,16 +333,13 @@ class KafkaConsumerTest extends TestCase
             [
                 new TopicPartition(KAFKA_TEST_TOPIC, 0),
             ],
-            (int) KAFKA_TEST_TIMEOUT_MS
+            KAFKA_TEST_TIMEOUT_MS
         );
 
         $this->assertCount(1, $topicPartitions);
         $this->assertSame(2, $topicPartitions[0]->getOffset());
     }
 
-    /**
-     * @medium
-     */
     public function testGetCommittedOffsets(): void
     {
         $conf = new Conf();
@@ -363,7 +354,7 @@ class KafkaConsumerTest extends TestCase
             [
                 new TopicPartition(KAFKA_TEST_TOPIC, 0),
             ],
-            (int) KAFKA_TEST_TIMEOUT_MS
+            KAFKA_TEST_TIMEOUT_MS
         );
 
         $this->assertCount(1, $topicPartitions);
@@ -371,7 +362,7 @@ class KafkaConsumerTest extends TestCase
 
         $consumed = 0;
         while ($consumed < 2) {
-            $message = $consumer->consume((int) KAFKA_TEST_TIMEOUT_MS);
+            $message = $consumer->consume(KAFKA_TEST_TIMEOUT_MS);
             if ($message->err === RD_KAFKA_RESP_ERR__TIMED_OUT) {
                 continue;
             }
@@ -383,7 +374,7 @@ class KafkaConsumerTest extends TestCase
             [
                 new TopicPartition(KAFKA_TEST_TOPIC, 0),
             ],
-            (int) KAFKA_TEST_TIMEOUT_MS
+            KAFKA_TEST_TIMEOUT_MS
         );
 
         $this->assertCount(1, $topicPartitions);
@@ -403,7 +394,7 @@ class KafkaConsumerTest extends TestCase
             [
                 new TopicPartition(KAFKA_TEST_TOPIC, 0, $future),
             ],
-            (int) KAFKA_TEST_TIMEOUT_MS
+            KAFKA_TEST_TIMEOUT_MS
         );
 
         $this->assertCount(1, $topicPartitions);
@@ -424,14 +415,14 @@ class KafkaConsumerTest extends TestCase
         $producerTopic = $producer->newTopic(KAFKA_TEST_TOPIC);
         $producerTopic->produce(RD_KAFKA_PARTITION_UA, 0, 'offsetsForTimes1');
         $producerTopic->produce(RD_KAFKA_PARTITION_UA, 0, 'offsetsForTimes2');
-        $producer->flush((int) KAFKA_TEST_TIMEOUT_MS);
+        $producer->flush(KAFKA_TEST_TIMEOUT_MS);
 
         $consumer = new KafkaConsumer($conf);
         $topicPartitions = $consumer->offsetsForTimes(
             [
                 new TopicPartition(KAFKA_TEST_TOPIC, 0, $nearNow),
             ],
-            (int) KAFKA_TEST_TIMEOUT_MS
+            KAFKA_TEST_TIMEOUT_MS
         );
 
         $this->assertCount(1, $topicPartitions);
@@ -451,7 +442,7 @@ class KafkaConsumerTest extends TestCase
             [
                 new TopicPartition(KAFKA_TEST_TOPIC, 0, $past),
             ],
-            (int) KAFKA_TEST_TIMEOUT_MS
+            KAFKA_TEST_TIMEOUT_MS
         );
 
         $this->assertCount(1, $topicPartitions);
@@ -466,7 +457,7 @@ class KafkaConsumerTest extends TestCase
 
         $consumer = new KafkaConsumer($conf);
 
-        $metadata = $consumer->getMetadata(true, null, (int) KAFKA_TEST_TIMEOUT_MS);
+        $metadata = $consumer->getMetadata(true, null, KAFKA_TEST_TIMEOUT_MS);
 
         $this->assertInstanceOf(Metadata::class, $metadata);
     }
