+++ tests/RdKafka/QueueTest.php
@@ -19,7 +19,7 @@ class QueueTest extends TestCase
         $producer->addBrokers(KAFKA_BROKERS);
         $producerTopic = $producer->newTopic(KAFKA_TEST_TOPIC);
         $producerTopic->produce(0, 0, __METHOD__);
-        $producer->flush((int) KAFKA_TEST_TIMEOUT_MS);
+        $producer->flush(KAFKA_TEST_TIMEOUT_MS);
 
         $consumer = new Consumer();
         $consumer->addBrokers(KAFKA_BROKERS);
@@ -29,7 +29,7 @@ class QueueTest extends TestCase
         $consumerTopic = $consumer->newTopic(KAFKA_TEST_TOPIC);
         $consumerTopic->consumeQueueStart(0, rd_kafka_offset_tail(1), $queue);
 
-        $message = $queue->consume((int) KAFKA_TEST_TIMEOUT_MS);
+        $message = $queue->consume(KAFKA_TEST_TIMEOUT_MS);
 
         $this->assertInstanceOf(Message::class, $message);
         $this->assertSame(__METHOD__, $message->payload);
@@ -49,7 +49,7 @@ class QueueTest extends TestCase
 
         $mainQueue = new Queue(Api::getFFI()->rd_kafka_queue_get_main($consumer->getCData()));
 
-        $event = $mainQueue->poll((int) KAFKA_TEST_TIMEOUT_MS);
+        $event = $mainQueue->poll(KAFKA_TEST_TIMEOUT_MS);
 
         $this->assertInstanceOf(Event::class, $event);
         $this->assertSame(4 /* RD_KAFKA_EVENT_LOG */, $event->type());
